Flesh out a bit more what the data store class looks like.
This is an unholy mess of a commit.  It is somewhat embarrassing.  But
I need to checkpoint, and I need to make this available to myself on
other machines.

In particular, I still need to figure out what the Series class really
looks like and how it deals with the key (date or time) and value
(double or string) types in some transparent way.